Experiment with staged messaging
Goal was to reduce number of messages types and number of messages used
at run time.

Issue seems to be that even though subscriptions is getting the updated
model before JS runs the wrong decoder get's used and so Message
delivery is delayed by an update cycle.

I have made things sorta work with a hack but it defeats the purpose.
End effect is also not exactly what I wanted. Some logic left to be
implemented for Drag end in this commit. Probably need to revist event
types to make this all work.